+++ elenasrc2/engine/win32/win32routines.cpp
@@ -119,7 +119,7 @@ uintptr_t SystemRoutineProvider :: ExpandHeap(void* allocPtr, int newSize)
    return !r ? 0 : (uintptr_t)allocPtr;
 }
 
-uintptr_t SystemRoutineProvider :: ExpandPerm(void* allocPtr, int newSize)
+uintptr_t SystemRoutineProvider :: ExpandPerm(void* allocPtr, size_t newSize)
 {
    // allocate
    LPVOID r = VirtualAlloc(allocPtr, newSize, MEM_COMMIT, PAGE_READWRITE);
