+++ elenasrc2/engine/win64/win64routines.cpp
@@ -109,7 +109,7 @@ uintptr_t SystemRoutineProvider::ExpandHeap(void* allocPtr, int newSize)
    return (uintptr_t)allocPtr;
 }
 
-uintptr_t SystemRoutineProvider::ExpandPerm(void* allocPtr, int newSize)
+uintptr_t SystemRoutineProvider::ExpandPerm(void* allocPtr, size_t newSize)
 {
    // allocate
    LPVOID r = VirtualAlloc(allocPtr, newSize, MEM_COMMIT, PAGE_READWRITE);
